requirements.txt
python-dotenv==1.0.1
selenium==4.17.2
webdriver-manager==4.0.1